+++ python/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2025-09-19
+
+### Changed
+- Removed verbose logging from fastmcp_patch for cleaner output
+
+## [0.2.3] - 2025-09-19
+
+### Fixed
+- Fixed well-known endpoint by using Starlette for response handling
+
 ## [0.2.1] - 2025-09-19
 
 ### Fixed